Made some changes to everything converting precision from int to double. Some conversions could not be done, so some of the methods changed to double return types had to be cast to int. Also I think I have completed XML_Problem, I also added a few changes to the GUI which allow arrows to be cleared and corrected the text on the arrow pop-up menu.

+++ src/moleculeMaker/Element.java
@@ -7,7 +7,7 @@ public class Element extends MoleculeComponent {
 	private int electrons;
 	private int charge;
 	
-	public Element(int x, int y)
+	public Element(double x, double y)
 	{
 		super();
 		
@@ -33,12 +33,12 @@ public class Element extends MoleculeComponent {
 	}
 
 	@Override
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 	
 	@Override
-	public int getY() {
+	public double getY() {
 		return y;
 	}
